requirements.txt
appdirs==1.4.4
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
numpy-stl==2.16.3
packaging==21.3
Pillow==8.4.0
pylas==0.4.3
pyopencl @ file:///C:/Users/pasha/Desktop/pyopencl-2021.2.10-cp38-cp38-win_amd64.whl
pyparsing==3.0.6
python-dateutil==2.8.2
python-utils==2.6.3
pytools==2021.2.9
six==1.16.0

pyopencl~=2021.2.10